backend/requirements.txt
Django==5.0.6
gunicorn==22.0.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-redis==5.4.0
requests==2.32.3
